requirements.txt
Django==2.0.2
djangorestframework==3.7.7
numpy==1.14.0
oauthlib==2.0.6
pandas==0.22.0
pymongo==3.6.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.4.3
requests-oauthlib==0.4.1
simplejson==3.13.2
six==1.7.3
tweepy==3.2.0
twitter==1.18.0
urllib3==1.22
